+++ dbmodel/sqlserver_db_setup_lcpserver.sql
@@ -6,7 +6,7 @@ CREATE TABLE content (
     id varchar(255) PRIMARY KEY NOT NULL,
     encryption_key varbinary(64) NOT NULL,
     location text NOT NULL,
-    length int,
+    length bigint,
     sha256 varchar(64),
     type varchar(255)
 );
@@ -17,8 +17,8 @@ CREATE TABLE license (
     provider varchar(255) NOT NULL,
     issued datetime NOT NULL,
     updated datetime DEFAULT NULL,
-    rights_print smallint DEFAULT NULL,
-    rights_copy smallint DEFAULT NULL,
+    rights_print bigint DEFAULT NULL,
+    rights_copy bigint DEFAULT NULL,
     rights_start datetime DEFAULT NULL,
     rights_end datetime DEFAULT NULL,
     content_fk varchar(255) NOT NULL,
